+++ EventListener/ThreadPermalinkListener.php
@@ -22,8 +22,8 @@ namespace FOS\CommentBundle\EventListener;
 
 use FOS\CommentBundle\Event\ThreadEvent;
 use FOS\CommentBundle\Events;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\HttpFoundation\RequestStack;
 
 /**
  * Responsible for setting a permalink for each new Thread object.
@@ -33,18 +33,16 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 class ThreadPermalinkListener implements EventSubscriberInterface
 {
     /**
-     * @var ContainerInterface
+     * @var RequestStack
      */
-    protected $container;
+    protected $requestStack;
 
     /**
-     * Constructor.
-     *
-     * @param ContainerInterface $container
+     * @param RequestStack $requestStack
      */
-    public function __construct(ContainerInterface $container)
+    public function __construct(RequestStack $requestStack)
     {
-        $this->container = $container;
+        $this->requestStack = $requestStack;
     }
 
     /**
@@ -54,12 +52,7 @@ class ThreadPermalinkListener implements EventSubscriberInterface
      */
     public function onThreadCreate(ThreadEvent $event)
     {
-        $request = null;
-        if ($this->container->has('request_stack')) {
-            $request = $this->container->get('request_stack')->getCurrentRequest();
-        } elseif (method_exists($this->container, 'isScopeActive') && $this->container->isScopeActive('request')) {
-            $request = $this->container->get('request');
-        }
+        $request = $this->requestStack->getCurrentRequest();
 
         if (null === $request) {
             return;